requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
aiofiles==23.2.1
tqdm==4.66.1
google-generativeai==0.3.1
python-dotenv==1.0.0
